Split elemental analysis test classes into separate modules
This allows them to run in parallel (at the process level)
within ctest.
It also avoids some strange interplay that was seen by the
tests on Windows running methods overlapping
with each other.
